feat: 로그아웃 구현 (#5)
* feat(auth): 로그아웃 구현

* fix(Auth): 인증 로직 수정

+++ src/main/java/com/tave/alarmissue/redis/service/RefreshTokenRedisService.java
@@ -29,8 +29,11 @@ public class RefreshTokenRedisService {
         return refreshTokenRepository.findById(userId);
     }
 
-    public void deleteRefreshToken(Long userId) {
-        refreshTokenRepository.deleteById(userId);
+    public boolean deleteRefreshToken(Long userId) {
+        if (refreshTokenRepository.existsById(userId)) {
+            refreshTokenRepository.deleteById(userId);
+            return true;
+        }
+        return false;
     }
-
 }
